+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## TBD
+## 5.8.0 (2021-09-15)
 
 * Address task dependency warning when using APK splits
   [#408](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/408)